backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
beautifulsoup4
requests
pydantic